pyproject.toml
[project]
name = "development-assistant"
version = "0.1.0"
description = "Development Assistant"
authors = [
    {name = "geekmaxi",email = "s.showling@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "langsmith[langsmith-pyo3] (>=0.3.8,<0.4.0)",
    "langchain (>=0.3.0,<0.4.0)",
    "langchain-cohere (>=0.3.0,<0.4.0)",
    "langchain-google-genai (>=2.0.0,<3.0.0)",
    "langchain-core (>=0.3.10,<0.4.0)",
    "langchain-community (>=0.3.0,<0.4.0)",
    "langchain-anthropic (>=0.2.0,<0.3.0)",
    "langchain-fireworks (>=0.2.0,<0.3.0)",
    "langchain-groq (>=0.2.0,<0.3.0)",
    "langchain-weaviate (>=0.0.3,<0.1.0)",
    "langgraph (>=0.2.32,<0.3.0)",
   
    "langchain-chroma (>=0.1.2)",
    "weaviate-client (>=4.10.4,<5.0.0)",
    "voyageai (>=0.3.2,<0.4.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "langserve (>=0.3.1,<0.4.0)",
    "sse-starlette (>=2.2.1,<3.0.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "einops (>=0.8.1,<0.9.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "grandalf (>=0.8,<0.9)",
    "langchain-openai (>=0.3.6,<0.4.0)",
    "langchain-deepseek (>=0.1.2,<0.2.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncmy (>=0.2.10,<0.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
